+++ app/src/main/java/nl/constantdynamics/everyday/EverydayApplication.kt
@@ -10,5 +10,6 @@ class EverydayApplication : Application() {
     override fun onCreate() {
         super.onCreate()
         container = AppContainer(this)
+        container.startOnderhoud()
     }
 }